+++ pslipstream/__init__.py
@@ -18,102 +18,120 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 
-# std
-import os
 import argparse
 import builtins as g
-import webbrowser
-import requests
 import hashlib
+import os
+import webbrowser
 
-# slipstream
+import requests
+from appdirs import user_data_dir
+try:
+    from cefpython3 import cefpython as cef
+except Exception as e:
+    if str(e).startswith("Python version not supported: 3.8"):
+        import site
+        raise ImportError(
+            "Oh no, cefpython is *still* not supporting 3.8 officially!! >:((\n"
+            f"Not to worry, just open '{site.getsitepackages()[0]}/cefpython3/__init__.py' and "
+            "change `elif sys.version_info[:2] == (3, 7):` (near ln60) to "
+            "`elif sys.version_info[:2] in [(3, 7), (3, 8)]:`\n"
+            "Once done, it will work fine on Python 3.8 :)"
+        )
+    raise
+
 import pslipstream.__version__ as meta
 from pslipstream.config import Config
+from pslipstream.dvd import Dvd
 from pslipstream.gui import Gui
 from pslipstream.helpers import get_device_list
-from pslipstream.dvd import Dvd
+from pslipstream.log import Log
+from pslipstream.progress import Progress
 
 
 def main():
+    # Initialize custom global variables
+    g.ARGS = get_arguments()
+    g.LOG = Log()  # Logger, everything written here gets print()'d and sent to GUI
+    g.PROGRESS = Progress()  # Progress Bar, controls only the GUI's progress bar.
+    g.DBG = g.ARGS.dbg  # Debug switch, enables debugging specific code and logging
+    g.CFG = Config(meta.__config_file__)
+    g.CFG.load()
 
-    # Parse Arguments
-    ArgParser = argparse.ArgumentParser()
-    ArgParser.add_argument(
+    # Print License if asked
+    if g.ARGS.license:
+        if not os.path.exists("LICENSE"):
+            # download the license from the official source if not found locally
+            lic_url = "https://www.gnu.org/licenses/gpl-3.0.txt"
+            lic_text = requests.get(lic_url).text
+            # noinspection SpellCheckingInspection
+            expected_md5 = "1ebbd3e34237af26da5dc08a4e440465"
+            computed_md5 = hashlib.md5(lic_text.encode("utf-8")).hexdigest()
+            if computed_md5 != expected_md5:
+                print(
+                    "License file was not found locally and it also failed to download :(\n"
+                    f"If you would like to read the LICENSE file, head over to {lic_url}"
+                )
+            else:
+                print(lic_text)
+        else:
+            with open("LICENSE", mode="rt", encoding="utf-8") as f:
+                print(f.read())
+        exit(0)
+
+    # Prepare Metadata
+    meta.__cef_version__ = cef.GetVersion()
+    meta.__user_dir__ = user_data_dir(meta.__title_pkg__, meta.__author__)
+    meta.__config_file__ = os.path.join(meta.__user_dir__, "config.yml")
+
+    # Get and Print Runtime Details
+    g.LOG.write(get_runtime_details() + "\n")
+
+    # Let's get to it
+    if g.ARGS.cli:
+        cli()
+    else:
+        gui()
+
+
+def get_arguments():
+    ap = argparse.ArgumentParser()
+    ap.add_argument(
         "--dev",
         nargs="*",
         required=False,
-        help="This isn't a debug mode, it's mainly for internally switching data, don't use this unless you know what they do!!",
+        help="This is only to be used by dev's that know what it does!!",
     )
-    ArgParser.add_argument(
+    ap.add_argument(
         "--dbg",
         action="store_true",
         default=False,
         required=False,
-        help="Debug mode, may have increased logging that could hurt performance, don't enable unless you debugging.",
+        help="Debug mode, may have increased logging that could hurt performance.",
     )
-    ArgParser.add_argument(
+    ap.add_argument(
         "--license",
         action="store_true",
         default=False,
         required=False,
         help="View license details",
     )
-    ArgParser.add_argument(
+    ap.add_argument(
         "--cli",
         action="store_true",
         default=False,
         required=False,
         help="Setting this stops the GUI from running",
     )
-    ArgParser.add_argument(
+    ap.add_argument(
         "-d",
         "--device",
         type=str,
         default="",
         required=False,
-        help="Choose device for backup (e.g. '/dev/sr0', '/mnt/dvdrw', 'E:')",
+        help="Choose device for backup (e.g. '/dev/sr0', '/mnt/dvd-rw', 'E:')",
     )
-    args = ArgParser.parse_args()
-
-    # Initialize custom global variables
-    # These need to be global to be accessed more easily
-    # throughout the code, otherwise id have to deal with messy
-    # imports and/or passing it from function to function
-    g.LOG = Log()  # Logger, everything written here gets print()'d and sent to GUI
-    g.PROG = Progress()  # Progress Bar, controls only the GUI's progress bar
-    g.DBG = args.dbg  # Debug switch, enables debugging specific code and logging
-    g.CFG = Config(meta.__config_file__)
-    g.CFG.load()
-
-    # Print License if asked
-    if args.license:
-        if not os.path.exists("LICENSE"):
-            # download the license from the official source if not found locally
-            md5 = hashlib.md5()
-            lic_url = "https://www.gnu.org/licenses/gpl-3.0.txt"
-            lic = requests.get(lic_url).text
-            md5.update(lic.encode("utf-8"))
-            if md5.hexdigest() != "1ebbd3e34237af26da5dc08a4e440465":
-                # this is just to ensure that the license file content doesn't change or
-                # fail to download or get re-routed or mitm'd e.t.c. It's important for a
-                # license to display correctly!
-                print(
-                    "License file was not found locally (pesky redistributors! >:L), but "
-                    "it also failed to download :(\nIf you would like to read the LICENSE"
-                    f", more information can be found here: {lic_url}"
-                )
-            else:
-                print(lic)
-        else:
-            with open("LICENSE", mode="rt", encoding="utf-8") as f:
-                print(f.read())
-        exit(0)
-
-    # Get and Print Runtime Details
-    g.LOG.write(get_runtime_details() + "\n")
-
-    # Let's get to it
-    (CLI_ROUTE if args.cli else GUI_ROUTE)(args)
+    return ap.parse_args()
 
 
 def get_runtime_details():
@@ -134,20 +152,19 @@ def get_runtime_details():
     )
 
 
-
-def GUI_ROUTE(args):
+def gui():
     # set ui index location based on environment
-    if args.dev:
+    if g.ARGS.dev:
         port = None
-        if "ui-build" in args.dev:
+        if "ui-build" in g.ARGS.dev:
             # gatsby build && gatsby serve
             port = 9000
-        elif "ui-develop" in args.dev:
+        elif "ui-develop" in g.ARGS.dev:
             # gatsby develop
             port = 8000
-        meta.__ui_index__ = f"http://localhost:{port}/"
+        meta.__ui_index__ = "http://localhost:" + str(port)
     else:
-        meta.__ui_index__ = "https://phoenix-slipstream.netlify.app/"
+        meta.__ui_index__ = "https://phoenix-slipstream.netlify.app"
     # create gui, and fire it up
     g.GUI = Gui(
         url=meta.__ui_index__,
@@ -172,46 +189,10 @@ def GUI_ROUTE(args):
     g.GUI.mainloop()
 
 
-def CLI_ROUTE(args):
-
-    dvd = Dvd()
-    dvd.open(args.device)
-    dvd.create_backup()
-
-
-class Log:
-    def __init__(self):
-        self.entries = []
-        self.js = None
-        self.max_entries = 100
-
-    def write(self, entry, echo=True):
-        self.entries.append(entry)
-        while len(self.entries) > self.max_entries:
-            self.entries.pop(0)
-        if echo:
-            print(entry.strip())
-        if self.js:
-            # update js log
-            self.read_all()
-    
-    def set_js(self, js):
-        self.js = js
-        self.read_all()  # get initial log
-
-    def read_all(self, js=None):
-        entries = "\n".join(self.entries).strip()
-        if self.js:
-            self.js.Call(entries)
-        return entries
-
-class Progress:
-    def __init__(self):
-        self.progress = 0
-        self.c = None
-    
-    def set_c(self, js):
-        self.c = js
+def cli():
+    d = Dvd()
+    d.open(g.ARGS.device)
+    d.create_backup()
 
 
 if __name__ == "__main__":