+++ pslipstream/progress.py
@@ -1,0 +1,8 @@
+class Progress:
+    def __init__(self):
+        self.progress = 0
+        self.c = None
+
+    def set_c(self, js):
+        # todo ; rename function to set_js_callback to be more descriptive
+        self.c = js